+++ src/settings/Models/RequestPolicy/RequestPolicy.js
@@ -10,6 +10,7 @@ export default class RequestPolicy {
     this.name = policy.name;
     this.description = policy.description;
     this.requestTypes = policy.requestTypes;
+    this.allowedServicePoints = policy.allowedServicePoints;
     this.metadata = new Metadata(policy.metadata);
   }
 }