+++ src/settings/RequestPolicy/RequestPolicySettings.test.js
@@ -10,6 +10,7 @@ import { EntryManager } from '@folio/stripes/smart-components';
 
 import RequestPolicySettings, {
   parseInitialValues,
+  findServicePointById,
 } from './RequestPolicySettings';
 import RequestPolicyDetail from './RequestPolicyDetail';
 import RequestPolicyForm from './RequestPolicyForm';
@@ -17,6 +18,7 @@ import RequestPolicy from '../Models/RequestPolicy';
 import normalize from './utils/normalize';
 
 import {
+  REQUEST_TYPE_RULES,
   requestPolicyTypes,
 } from '../../constants';
 
@@ -89,7 +91,7 @@ describe('RequestPolicySettings', () => {
 
     expect(EntryManager).toHaveBeenCalledWith(expect.objectContaining({
       parentMutator: mockedMutator,
-      parseInitialValues,
+      parseInitialValues: expect.any(Function),
       parentResources: defaultProps.resources,
       entryList: sortedEntyList,
       resourceKey: 'requestPolicies',
@@ -179,7 +181,38 @@ describe('RequestPolicySettings', () => {
   });
 });
 
+describe('findServicePointById', () => {
+  const id = 'id';
+  const servicePoints = {
+    records: [{ id }],
+  };
+
+  it('should return found record', () => {
+    expect(findServicePointById(servicePoints, id)).toEqual(servicePoints.records[0]);
+  });
+
+  it('should return undefined', () => {
+    expect(findServicePointById(servicePoints, 'testId')).toBeUndefined();
+  });
+});
+
 describe('parseInitialValues', () => {
+  const id = 'id';
+  const name = 'name';
+  const servicePoints = {
+    records: [
+      {
+        id,
+        name,
+      }
+    ],
+  };
+  const requestTypesRules = requestPolicyTypes.reduce((acc, requestType) => {
+    acc[requestType] = REQUEST_TYPE_RULES.ALLOW_ALL;
+
+    return acc;
+  }, {});
+
   it('should correctly process if "values" was not passed', () => {
     const expectedResult = {
       requestTypes: [
@@ -187,9 +220,11 @@ describe('parseInitialValues', () => {
         false,
         false,
       ],
+      allowedServicePoints: {},
+      requestTypesRules,
     };
 
-    expect(parseInitialValues()).toEqual(expectedResult);
+    expect(parseInitialValues(servicePoints)()).toEqual(expectedResult);
   });
 
   it('should correctly process passed "values" with "requestTypes"', () => {
@@ -203,8 +238,40 @@ describe('parseInitialValues', () => {
         false,
         true,
       ],
+      allowedServicePoints: {},
+      requestTypesRules,
     };
 
-    expect(parseInitialValues(mockedValues)).toEqual(expectedResult);
+    expect(parseInitialValues(servicePoints)(mockedValues)).toEqual(expectedResult);
+  });
+
+  it('should return correct data if only some service points allowed', () => {
+    const mockedValues = {
+      requestTypes: [requestPolicyTypes[0]],
+      allowedServicePoints: {
+        [requestPolicyTypes[0]]: [id],
+      },
+    };
+    const expectedResult = {
+      requestTypes: [
+        true,
+        false,
+        false,
+      ],
+      requestTypesRules: {
+        ...requestTypesRules,
+        [requestPolicyTypes[0]]: REQUEST_TYPE_RULES.ALLOW_SOME,
+      },
+      allowedServicePoints: {
+        [requestPolicyTypes[0]]: [
+          {
+            label: name,
+            value: id,
+          }
+        ],
+      },
+    };
+
+    expect(parseInitialValues(servicePoints)(mockedValues)).toEqual(expectedResult);
   });
 });