+++ src/settings/RequestPolicy/components/EditSections/GeneralSection/GeneralSection.test.js
@@ -3,6 +3,7 @@ import {
   render,
   screen,
   within,
+  cleanup,
 } from '@testing-library/react';
 
 import '../../../../../../test/jest/__mock__';
@@ -14,12 +15,19 @@ import {
   TextArea,
   TextField,
   Checkbox,
+  RadioButton,
+  MultiSelection,
 } from '@folio/stripes/components';
 import { Metadata } from '../../../../components';
 import GeneralSection, {
   renderTypes,
+  getDataOptions,
+  validateServicePointsList,
 } from './GeneralSection';
-import { requestPolicyTypes } from '../../../../../constants';
+import {
+  REQUEST_TYPE_RULES,
+  requestPolicyTypes,
+} from '../../../../../constants';
 
 jest.mock('../../../../components', () => ({
   Metadata: jest.fn(() => null),
@@ -36,6 +44,8 @@ describe('GeneralSection', () => {
     accordion: 'ui-circulation.settings.requestPolicy.generalInformation',
     name: 'ui-circulation.settings.requestPolicy.policyName',
     description: 'ui-circulation.settings.requestPolicy.policyDescription',
+    allowAllServicePoints: 'ui-circulation.settings.requestPolicy.requestTypes.allowAll',
+    allowSomeServicePoints: 'ui-circulation.settings.requestPolicy.requestTypes.allowSome',
   };
   const fieldCallOrderByPlace = {
     name: 1,
@@ -47,17 +57,32 @@ describe('GeneralSection', () => {
   const mockedConnect = jest.fn();
   const mockedValidateName = jest.fn();
   const getById = (id) => within(screen.getByTestId(id));
+  const renderTypesProps = {
+    handleChangeRequestTypesRules: jest.fn(),
+    requestTypesRules: {
+      Hold: REQUEST_TYPE_RULES.ALLOW_ALL,
+    },
+    servicePoints: [
+      {
+        id: 'id',
+        name: 'name',
+      }
+    ],
+    isLoading: false,
+  };
   const defaultProps = {
     isOpen: true,
     metadata: mockedMetadata,
     connect: mockedConnect,
     validateName: mockedValidateName,
+    ...renderTypesProps,
   };
 
   afterEach(() => {
     Metadata.mockClear();
     Field.mockClear();
     FieldArray.mockClear();
+    cleanup();
   });
 
   describe('when "isOpen" prop is true', () => {
@@ -115,6 +140,7 @@ describe('GeneralSection', () => {
       const expectedResult = {
         name: 'requestTypes',
         component: renderTypes,
+        ...renderTypesProps,
       };
 
       expect(FieldArray).toHaveBeenLastCalledWith(expectedResult, {});
@@ -143,7 +169,6 @@ describe('GeneralSection', () => {
   describe('renderTypes', () => {
     const policyTypesLabelId = 'ui-circulation.settings.requestPolicy.policyTypes';
     const polycyTypeTest = (name, index) => {
-      const number = index + 1;
       const expectedResult = {
         component: Checkbox,
         type: 'checkbox',
@@ -152,21 +177,136 @@ describe('GeneralSection', () => {
         name: `requestTypes[${index}]`,
       };
 
-      it(`should render ${number} policy with right props`, () => {
-        expect(Field).toHaveBeenNthCalledWith(number, expectedResult, {});
+      it(`should render ${name} policy with correct props`, () => {
+        expect(Field).toHaveBeenCalledWith(expectedResult, {});
       });
     };
 
-    beforeEach(() => {
-      render(
-        renderTypes()
-      );
+    describe('when all service points allowed', () => {
+      beforeEach(() => {
+        render(
+          renderTypes({
+            ...renderTypesProps,
+            fields: {
+              value: [true],
+            },
+          })
+        );
+      });
+
+      it('should render main label', () => {
+        expect(getById(testIds.policyTypes).getByText(policyTypesLabelId)).toBeVisible();
+      });
+
+      it('should render allow all service points label', () => {
+        const allowAllServicePointsLabel = screen.getByText(labelIds.allowAllServicePoints);
+
+        expect(allowAllServicePointsLabel).toBeInTheDocument();
+      });
+
+      it('should render allow some service points label', () => {
+        const allowSomeServicePointsLabel = screen.getByText(labelIds.allowSomeServicePoints);
+
+        expect(allowSomeServicePointsLabel).toBeInTheDocument();
+      });
+
+      it('should trigger allow all service points Field with correct props', () => {
+        const expectedProps = {
+          type: 'radio',
+          name: `requestTypesRules.${requestPolicyTypes[0]}`,
+          id: `${requestPolicyTypes[0]}AllowAllRadioButton`,
+          component: RadioButton,
+          value: REQUEST_TYPE_RULES.ALLOW_ALL,
+          onChange: expect.any(Function),
+          disabled: false,
+        };
+
+        expect(Field).toHaveBeenCalledWith(expect.objectContaining(expectedProps), {});
+      });
+
+      it('should trigger allow some service points Field with correct props', () => {
+        const expectedProps = {
+          type: 'radio',
+          name: `requestTypesRules.${requestPolicyTypes[0]}`,
+          id: `${requestPolicyTypes[0]}AllowSomeRadioButton`,
+          component: RadioButton,
+          value: REQUEST_TYPE_RULES.ALLOW_SOME,
+          onChange: expect.any(Function),
+          disabled: false,
+        };
+
+        expect(Field).toHaveBeenCalledWith(expect.objectContaining(expectedProps), {});
+      });
+
+      requestPolicyTypes.forEach(polycyTypeTest);
     });
 
-    it('should render main label', () => {
-      expect(getById(testIds.policyTypes).getByText(policyTypesLabelId)).toBeVisible();
+    describe('when some service points allowed', () => {
+      beforeEach(() => {
+        render(
+          renderTypes({
+            ...renderTypesProps,
+            requestTypesRules: {
+              Hold: REQUEST_TYPE_RULES.ALLOW_SOME,
+            },
+            fields: {
+              value: [true],
+            },
+          })
+        );
+      });
+
+      it('should trigger Field with correct props', () => {
+        const expectedProps = {
+          component: MultiSelection,
+          name: `allowedServicePoints.${requestPolicyTypes[0]}`,
+          id: `${requestPolicyTypes[0]}MultiSelect`,
+          validate: validateServicePointsList,
+          disabled: false,
+        };
+
+        expect(Field).toHaveBeenCalledWith(expect.objectContaining(expectedProps), {});
+      });
+    });
+  });
+
+  describe('getDataOptions', () => {
+    it('should correctly modify service points', () => {
+      const name = 'name';
+      const id = 'id';
+      const servicePoints = [
+        {
+          name,
+          id,
+          test: 'test',
+        }
+      ];
+      const expectedResult = [
+        {
+          label: name,
+          value: id,
+        }
+      ];
+
+      expect(getDataOptions(servicePoints)).toEqual(expectedResult);
     });
 
-    requestPolicyTypes.forEach(polycyTypeTest);
+    it('should return empty array', () => {
+      expect(getDataOptions()).toEqual([]);
+    });
+  });
+
+  describe('validateServicePointsList', () => {
+    it('should return undefined', () => {
+      const servicePoints = [{}];
+
+      expect(validateServicePointsList(servicePoints)).toBeUndefined();
+    });
+
+    it('should return an error', () => {
+      const servicePoints = [];
+
+      expect(validateServicePointsList(servicePoints)).toBeTruthy();
+    });
   });
 });